Remove redundant mgenc code
